# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Social Trend App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type QuestionCategory =
  (typeof QuestionCategory)[keyof typeof QuestionCategory];

export const QuestionCategory = {
  Dating: "Dating",
  Marriage: "Marriage",
  Gender: "Gender",
  Money: "Money",
  Values: "Values",
  Career: "Career",
} as const;

export type QuestionType = (typeof QuestionType)[keyof typeof QuestionType];

export const QuestionType = {
  single_choice: "single_choice",
  slider: "slider",
} as const;

export type QuestionStatus =
  (typeof QuestionStatus)[keyof typeof QuestionStatus];

export const QuestionStatus = {
  active: "active",
  draft: "draft",
} as const;

export interface Question {
  id: string;
  category: QuestionCategory;
  type: QuestionType;
  prompt: string;
  options: string[];
  status: QuestionStatus;
  responseCount: number;
}

export interface DistributionItem {
  option: string;
  percentage: number;
  count: number;
}

export interface SegmentResult {
  groupName: string;
  distribution: DistributionItem[];
}

export interface AggregatedResult {
  questionId: string;
  totalResponses: number;
  majorityAnswer: string;
  distribution: DistributionItem[];
  segments: SegmentResult[];
}

export interface SubmitResponseBody {
  sessionId: string;
  questionId: string;
  answer: string;
  predictedMajority: string;
}

export interface UserResponse {
  id: string;
  sessionId: string;
  questionId: string;
  answer: string;
  predictedMajority: string;
  createdAt: string;
}

export interface UserProfile {
  sessionId: string;
  ageRange?: string | null;
  gender?: string | null;
  region?: string | null;
  relationshipStatus?: string | null;
  answeredCount: number;
  predictionAccuracy: number;
  favoriteCategory?: string | null;
  badge?: string | null;
}

export interface UpdateDemographicsBody {
  sessionId: string;
  ageRange?: string | null;
  gender?: string | null;
  region?: string | null;
  relationshipStatus?: string | null;
}

export interface AppStats {
  totalResponses: number;
  totalQuestions: number;
  activeUsers: number;
  topCategory: string;
}

export type ListQuestionsParams = {
  category?: string;
};

export type GetProfileParams = {
  sessionId: string;
};