# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Social Trend App API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all active questions
 */
export const ListQuestionsQueryParams = zod.object({
  category: zod.coerce.string().optional(),
});

export const ListQuestionsResponseItem = zod.object({
  id: zod.string(),
  category: zod.enum([
    "Dating",
    "Marriage",
    "Gender",
    "Money",
    "Values",
    "Career",
  ]),
  type: zod.enum(["single_choice", "slider"]),
  prompt: zod.string(),
  options: zod.array(zod.string()),
  status: zod.enum(["active", "draft"]),
  responseCount: zod.number(),
});
export const ListQuestionsResponse = zod.array(ListQuestionsResponseItem);

/**
 * @summary Get today's featured question
 */
export const GetTodayQuestionResponse = zod.object({
  id: zod.string(),
  category: zod.enum([
    "Dating",
    "Marriage",
    "Gender",
    "Money",
    "Values",
    "Career",
  ]),
  type: zod.enum(["single_choice", "slider"]),
  prompt: zod.string(),
  options: zod.array(zod.string()),
  status: zod.enum(["active", "draft"]),
  responseCount: zod.number(),
});

/**
 * @summary Get a single question by ID
 */
export const GetQuestionParams = zod.object({
  id: zod.coerce.string(),
});

export const GetQuestionResponse = zod.object({
  id: zod.string(),
  category: zod.enum([
    "Dating",
    "Marriage",
    "Gender",
    "Money",
    "Values",
    "Career",
  ]),
  type: zod.enum(["single_choice", "slider"]),
  prompt: zod.string(),
  options: zod.array(zod.string()),
  status: zod.enum(["active", "draft"]),
  responseCount: zod.number(),
});

/**
 * @summary Get aggregated results for a question
 */
export const GetQuestionResultsParams = zod.object({
  id: zod.coerce.string(),
});

export const GetQuestionResultsResponse = zod.object({
  questionId: zod.string(),
  totalResponses: zod.number(),
  majorityAnswer: zod.string(),
  distribution: zod.array(
    zod.object({
      option: zod.string(),
      percentage: zod.number(),
      count: zod.number(),
    }),
  ),
  segments: zod.array(
    zod.object({
      groupName: zod.string(),
      distribution: zod.array(
        zod.object({
          option: zod.string(),
          percentage: zod.number(),
          count: zod.number(),
        }),
      ),
    }),
  ),
});

/**
 * @summary Submit a user answer and prediction
 */
export const SubmitResponseBody = zod.object({
  sessionId: zod.string(),
  questionId: zod.string(),
  answer: zod.string(),
  predictedMajority: zod.string(),
});

/**
 * @summary Get user profile stats
 */
export const GetProfileQueryParams = zod.object({
  sessionId: zod.coerce.string(),
});

export const GetProfileResponse = zod.object({
  sessionId: zod.string(),
  ageRange: zod.string().nullish(),
  gender: zod.string().nullish(),
  region: zod.string().nullish(),
  relationshipStatus: zod.string().nullish(),
  answeredCount: zod.number(),
  predictionAccuracy: zod.number(),
  favoriteCategory: zod.string().nullish(),
  badge: zod.string().nullish(),
});

/**
 * @summary Save user demographics
 */
export const UpdateDemographicsBody = zod.object({
  sessionId: zod.string(),
  ageRange: zod.string().nullish(),
  gender: zod.string().nullish(),
  region: zod.string().nullish(),
  relationshipStatus: zod.string().nullish(),
});

export const UpdateDemographicsResponse = zod.object({
  sessionId: zod.string(),
  ageRange: zod.string().nullish(),
  gender: zod.string().nullish(),
  region: zod.string().nullish(),
  relationshipStatus: zod.string().nullish(),
  answeredCount: zod.number(),
  predictionAccuracy: zod.number(),
  favoriteCategory: zod.string().nullish(),
  badge: zod.string().nullish(),
});

/**
 * @summary Get overall app stats (total responses, questions answered, etc.)
 */
export const GetStatsResponse = zod.object({
  totalResponses: zod.number(),
  totalQuestions: zod.number(),
  activeUsers: zod.number(),
  topCategory: zod.string(),
});